# Conflicts:
#	app/src/main/java/a75f/io/renatus/modbus/ModbusConfigViewModel.kt
#	data/build.gradle
#	domain/src/main/java/a75f/io/domain/logic/DefaultEquipBuilder.kt
#	domain/src/main/java/a75f/io/domain/logic/EquipBuilder.kt
#	domain/src/main/java/a75f/io/domain/logic/ProfileEquipBuilder.kt
#	domain/src/main/java/a75f/io/domain/logic/TunerEquipBuilder.kt
#	domain/src/main/java/a75f/io/domain/migration/MigrationHandler.kt
#	logic/src/main/java/a75f/io/logic/Globals.java

+++ app/src/main/java/a75f/io/renatus/schedules/NamedScheduleUnoccupiedDailog.java
@@ -61,7 +61,7 @@ public class NamedScheduleUnoccupiedDailog extends DialogFragment {
 
         title.setText(DAYS.get(getArguments().getInt(PARAM_DAY))+" | ");
         if(isCelsiusTunerAvailableStatus()){
-            unOccupied.setText(UnitUtils.fahrenheitToCelsiusRelative(schedule.getUnoccupiedZoneSetback()) + "\u00B0C");
+            unOccupied.setText((int)UnitUtils.fahrenheitToCelsiusRelative(schedule.getUnoccupiedZoneSetback()) + "\u00B0C");
         }else {
             unOccupied.setText(schedule.getUnoccupiedZoneSetback().toString() + "\u00B0F");
         }
